Add final address pass-through
Add circuit to pass-through the final address signal. Be careful -- this
signal is passed out with the accumulator total and not the difference
pixel it is associated with. This design decision was made since there
was no obvious reason why the difference pixel needs to know it is the
last pixel while there is reason to want to know when the accumulator
total has updated.